Limit Condition Attributes to one attribute type per condition
