+++ condition-web/src/components/ConditionDetails/ConditionAttribute/Constants.ts
